hot fix of mean function

+++ simple_math.c
@@ -7,7 +7,7 @@
 #include "simple_math.h"
 
 float mean(int a, int b) {
-	return (a+b)/2;
+	return (float)(a+b)/2.;
 }
 
 int max(int a, int b) {
